moved MotorManager to a diffrent package
