+++ src/main/java/com/basicMotor/configuration/BasicMotorConfig.java
@@ -1,11 +1,12 @@
 package com.basicMotor.configuration;
 
 import com.basicMotor.BasicMotor;
+import com.basicMotor.Manager.MotorManagerConfig;
 import com.basicMotor.gains.ControllerConstrains;
 import com.basicMotor.gains.ControllerFeedForwards;
 import com.basicMotor.gains.ControllerGains;
 import com.basicMotor.gains.PIDGains;
-import com.basicMotor.MotorManager;
+import com.basicMotor.Manager.MotorManager;
 import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.math.trajectory.TrapezoidProfile;
 import java.util.function.Function;
