+++ src/main/java/com/basicMotor/gains/ControllerConstrains.java
@@ -1,6 +1,6 @@
 package com.basicMotor.gains;
 
-import com.basicMotor.MotorManager;
+import com.basicMotor.Manager.MotorManager;
 import com.basicMotor.controllers.Controller;
 import com.basicMotor.measurements.Measurements;
 import edu.wpi.first.math.MathUtil;
